Small code style change
When there's a single non-error ("happy") execution path in the code,
it's usually clearer to lift that out of a nested block and guard it
with conditionals.  The error paths come first, but the happy path is
still emphasized more by being unindented.

On the other hand, when there's more than one happy path, it's usually
clearer to push them all into the same mutually-exclusive conditional
and put error paths last.  This emphasizes the variety of happy paths
and that there isn't a single one.