Add absolute path to custom CSS for production
The app is deployed at /switchboard in production, and Datasette doesn't
include a way to adjust the extra CSS URLs for different deployment vs.
dev paths (without using entirely separate configs).

By including both URLs, we guarantee that one will work in dev and the
other in production.  (And it's idempotent, so if both work one day,
it's not a problem.)

Resolves #7.