Fix broken links due to Datasette bug using client-side rewrites
A small piece of JS added to each page looks for such broken links by
their incorrect origin and rewrites them.

The reason for including two (seemingly duplicate) extra JS URLs is
discussed in "Add absolute path to custom CSS for production" (4342a46).

I wrote a quick reverse proxy in order to test locally behind an actual
proxy.  It uses Perl's excellent Plack library, because that was handy
and I couldn't find something similar for Python.  It's committed here
in case it's useful in the future, but I don't expect us to run it
often.

Resolves #8.