Reduce number of simultaneous connections to REDCap
We started seeing consistent TimeoutErrors thrown in production.  This
happened after a major REDCap upgrade, but it's not clear if it's
related on not.

The timeouts appear to be arising from the OS-level networking stack
since we don't set an explicit timeout for the "requests" library and
the exception isn't the Python-level socket library timeout class
socket.timeout.

On a hunch that REDCap can no longer service as many connections
simultaneously, reduce from 15 → 8.  Since we currently have 16
projects, each thread will take 2 projects.