Deep link to the PCDEQC form where applicable
Instead of dropping users into the REDCap record home page on any
barcode lookup, if it's a swab-and-send barcode, drop them into the
specific instance's post collection data entry quality control form.

If the PCDEQC form does not exist (e.g. in the SCAN In-Person
Enrollments project) or the barcode is a kiosk barcode, then continue
redirecting users to the record home page.