Account for repeat instance in duplicate_record_ids
Add REDCap repeat instance as a grouping key to the
`duplicate_record_ids` table in Datasette. We now have a project with
repeating instruments, so false positives for duplicate record IDs
(that had distinct repeat instance IDs) had inundated the table.

We were incorrectly referring to the REDCap unique event name as
"record_arm" in the barcode database; correct this.