Fix: redirect to PCDEQC form only if it exists
The PCDEQC form does not exist in any SCAN In-Person enrollment records,
despite having the `post_data_collection_qc_complete` field for every
record in REDCap. We can utilize the fact that for records where this
form doesn't exist, the value for `post_data_collection_qc_complete` is
an empty string (versus "0" if the form exists but is not yet complete).