Add timeout to SQL command
We are seeing instances where datasette is locking the sqlite database
and causing the `derived-tables.sql` task to fail with the error
`database is locked`.
This commit adds a 10 second timeout to the top of the SQL file.
This will let the command wait up to 10 seconds for the lock to
get removed so that it can run.

To reproduce the error run this batch:
for i in {1..100}
do
   sqlite3 data/sfs-redcap.sqlite < derived-tables.sql
done

And while that's running use datasette to access a barcode like:
http://127.0.0.1:3002/sfs-redcap/lookup-barcode?barcode=aaaaaaaa
Run that query a few times in a row and without the timeout you'll see
the `database is locked` error.