Add datasette timeout
Switchboard users have reported getting a
`TypeError: Cannot read property 'length' of undefined`
error when trying to look up a barcode. This is an intermittent
error, and we tracked this back to to the sqlite database being
locked while records are loading.

This commit adds a timeout to datasette config to let a query
try for 10 seconds (instead of the default of 1 second). This should
accommodate periods when the query needs to wait for the database
to be available.

In here we also look for the specific error situation and show a
clearer exception message.