Ditch Pipenv for pip-tools
Pipenv was nice at first for all our projects, but with time has shown
its warts with dependency-updating behaviour that's hard to reason about
and slowness or outright failure to resolve dependencies.

Using this project as an example and trial case for other projects,
switch to using pip-tools to manage a requirements.in file which is
compiled to a fully-specified requirements.txt file.  The
requirements.txt file is then installed into a vanilla virtualenv
located in .venv/.

There are some rough edges in pip-tools, notably around VCS dependency
support, but pip-tools is being actively developed and is tracking these
issues.

The versions of our deps are constrained to the same versions as were in
Pipfile.lock to avoid upgrading them with this change.  We can remove or
update the constraints in the future to upgrade them.