Include barcodes from AFH and Workplace Outbreaks
Include barcodes from Adult Family Homes and Workplace Outbreaks on the
switchboard. This project is a "classic" RedCAP project, such that it is not
longitudinal and does not have events or arms. The RedCAP API call to fetch
events will fail for classic projects, and we don't actually need to provide
and arm or event to construct the deep link URL into the PCDEQC instrument for
records in this classic project. Include barcode variables
(core_collection_barcode and return_collection_barcode) from the new project.