Enable SQLite WAL mode in the Makefile
Explicitly set the SQLite database to the WAL mode. This handles
the scenario when we delete the database files and let the code create
new ones after we modify the columns in the underlying tables. Without
explicitly setting WAL mode, the database would stay in the 'delete'
(the default) mode.

Setting WAL mode when the database is already in WAL mode in a no-op.