Set the database file as group-writable
SQLite defaults to creating databases as group and world readable but
writable only by the user (u=rw,go=r).¹  We can't reasonably change that
behaviour since it is a compile-time option, so instead we need to chmod
it ourselves after creation.  The added call isn't conditional on
creation, but it is harmless to chmod it anyway.

If the main database file isn't group-writable, then the sidecar files
(.wal and .shm) for WAL journaling also aren't marked group-writable.
This is problematic as the sidecar files are automatically
created/deleted by SQLite and thus may be created by a different user
than the owner of the main database file, rendering the two sets of
files writable only by different users.  All SQLite clients need to be
able write to the sidecar files for WAL journaling to work.  We use
POSIX ACLs to grant proper default permissions, but these are altered to
a reduced effective set of perms by the creation mode SQLite uses.
SQLite will create the sidecar files with a mode matching the main
database file.

¹ https://www.sqlite.org/compile.html#default_file_permissions