fix-links: Only adjust path if it's incorrect
While the host is still wrong, Datasette now includes the right base_url
path (because we now make the backend request to that same path), so
only add the path if it's missing it.  Though I haven't observed this
case yet, it's simple to support just in case Datasette still produces
the odd link or two without the base_url path.