Index all the barcode columns
This dramatically improves barcode dialing speed again now that we have
much more data.

It'd be nice to not repeat this list of column names everywhere, but
that's a task for another day.  And it will be moot if we move to a
single-column data model, as Chris is trialing.