Generate record link JSON at query time
The CSV DictWriter was writing the JSON for the `record_link` column
using a single quote instead of a double quote around the JSON key names
and values. This was causing a `json.loads` error in html_links.py.

With this change, we don't store the JSON in the table. Instead,
at query time, the SQL concatenates to build the JSON. The two components
that go into the JSON are `record_url` which was already a column
and a new column named `record_link_label`.