Pivot record_barcodes table to use one barcode column
In this commit, individual REDCap barcode columns are replaced by a
single barcode column and a redcap_field column to capture the REDCap
field that held the barcode value. This "pivots" the record_barcodes
table from wide to tall, with one barcode per row. There are 3 key
advantages to this approach:
1. The SQL to find a record containing a barcode is greatly simplified.
There's no longer a big `coalesce` that looks through many columns;
instead it looks at only one column.
2. When new REDCap barcode fields need to be added, the SQL won't
have to be modified. The new fields will just need to be incorporated into
`export-record-barcodes`.
3. The user experience is improved for the scenario when a user
dials a barcode and that returns more than one result. Instead of having
to scroll far to the right to see (and try to keep in mind) all the
columns, the user will see the fields that contain the barcode in
real estate that doesn't require scrolling.

Because this pivot takes away the view that shows all the barcodes in all
the columns, the `see-related-barcodes` query has been added to show all
the barcodes that belong to the REDCap record(s) to which the input
barcode belongs. This view offers 2 key advantages over the current design:
1. It shows all the barcodes related to the REDCap records, not just all
the barcodes for the particular REDCap instance that the input barcode
belongs to.
2. The view is always available. A Switchboard user can always plug in a
barcode and see its related barcodes. Currently, the similar view is only
available if dialing a barcode returns more than one result.

This commit adds a `unique_record_id` column to the record_barcodes table.
This gives a single column to uniquely identify the REDCap record that a
barcode came from. Having this as a single column simplifies some SQL a bit.
Currently, this is populated with the REDCap project ID |
REDCap record ID. In the event that the group uses multiple REDCap
instances in the future, the REDCap project API URL should be added
to the unique identifier.

The `derived-tables.sql` SQL has been updated to show occurences where the
record_barcodes table contains more than one row for the same REDCap
barcode field for a particular REDCap record instance. This would
indicate a fault in the data refresh. For example, the existing table
might not have gotten dropped before fresh data was added.