Disable REDCap Project 34101 from Switchboard
Overnight, the API token being used for this project on the backoffice server became invalid and is affecting the Switchboard refresh job. A majority of the team have lost access to the project. As a temporary fix, the project is commented out of the Switchboard pipeline. This project does not receive any new samples and is currently on cleanup/analysis on REDCap.