Remove Yakima Studies from Project List
We recently lost access to the Yakima study (REDCap PID 34101) and will not be getting our access back, so are removing it from the project list. Also included this change is the removal of the Spanish Pilot version of this study (PID 34701) which we still have access to but will likely lose in the future. This change will preempt those run failures for the Spanish project.